+++ src/mods/scala/draco/DrakeCLI.scala
@@ -123,8 +123,6 @@ object DrakeCLI {
           a => Json.fromValues(a.map(carried)),
           o => Json.fromJsonObject(io.circe.JsonObject.fromIterable(o.toIterable.map {
             case ("value", v) => "value" -> Json.fromString(Drake.defaultValue(Drake.expression(v)))
-            case (k, v) if k == "derivation" || k == "modules" =>
-              k -> Json.fromValues(v.asArray.getOrElse(Vector.empty).map(_.mapObject(_.remove("namePackage"))))
             case (k, v) => k -> carried(v)
           })))
         if (carried(TypeDefinition.encoder(Drake.parse(rendered))) != carried(TypeDefinition.encoder(td))) {